+++ lib/services/qiita-service.js
@@ -1,6 +1,6 @@
 'use babel'
 
-import ServiceBase from './service_base'
+import ServiceBase from './service-base'
 
 export default class QiitaService extends ServiceBase {
   constructor(accessToken) {
