+++ lib/services/service-base.js
